pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pydata-iceberg"
version = "0.1.0"
description = 'A demo of Iceberg and friends'
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Anders Bogsnes", email = "andersbognes@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "pyiceberg",
  "psycopg2-binary",
  "ibis-framework[duckdb,polars,datafusion]",
  "httpx",
  "polars",
  "pandas",
  "pyarrow",
  "s3fs",
  "fsspec",
  "pyathena[arrow,sqlalchemy]",
  "kaggle",
  "minio",
  "pydantic-settings",
  "pydantic",
]

[project.scripts]
demo = "pydata_iceberg.cli:cli"

[tool.coverage.run]
source_pkgs = ["pydata_iceberg", "tests"]
branch = true
parallel = true
omit = [
  "src/pydata_iceberg/__about__.py",
]

[tool.coverage.paths]
pydata_iceberg = ["src/pydata_iceberg", "*/pydata-iceberg/src/pydata_iceberg"]
tests = ["tests", "*/pydata-iceberg/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
